gbank_backend/requirements.txt
flask
google-cloud-firestore
flask_cors
google-auth
